+++ R/cpmr-package.R
@@ -1,0 +1,7 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @import Rfast
+## usethis namespace: end
+NULL