+++ R/zzz.R
@@ -1,2 +1,0 @@
-#' @import Rfast
-NULL